Move `AccountPage` to pages folder
